WIP for text widget -> text actor refactor
Due to some issues with how rotation was being handled in the text widget,
I've ripped out the text widget and am using raw text actors now.

I had to add a second renderer to the manager to place actors on top of widgets,
which is still a little buggy (after hiding and showing, widgets are on top... wat)

Text actors are, at this point, very nearly placed exactly where they should be, which is
something the previous implementation never had going for it.